+++ includes/tiff_helpers.h
@@ -17,7 +17,7 @@ void tiff_read_to_color_buffer(TIFF *tif, uint32_t *buffer, uint32_t height, uin
  * @param tif               pointer to TIFF struct
  * @param buffer            
  */
-void tiff_read_to_intensity_buffer(TIFF *tif, float *buffer, uint32_t height, uint32_t width);
+void tiff_read_to_intensity_buffer(TIFF *tif, double *buffer, uint32_t height, uint32_t width);
 
 /**
  * Returns height of TIFF
